@@ -1,3 +1,3 @@
 module Rails3JQueryAutocomplete
-  VERSION = '1.0.11'
+  VERSION = '1.0.14'
 end
